The jini jars were renamed from foo-2.1.jar to foo.jar -- the version information in the jar name was breaking the ClassServer.
Refactored jini and icu jars into subdirectories for easier packaging.

Cleaned up the service test suites and included an embedded ClassServer so that downloadable code is always available during service tests.

Added IRawStore#delete()


git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/trunk@240 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata/src/java/com/bigdata/btree/IndexSegmentFileStore.java
@@ -313,15 +313,25 @@ public class IndexSegmentFileStore extends AbstractRawStore implements IRawStore
 
     }
     
-    public void closeAndDelete() {
-        
-        close();
+    public void delete() {
+
+        if (open)
+            throw new IllegalStateException();
         
         if(!file.delete()) {
             
-            System.err.println("WARN: Could not delete: "+file.getAbsolutePath());
+            throw new RuntimeException("Could not delete: "
+                    + file.getAbsolutePath());
             
         }
+
+    }
+    
+    public void closeAndDelete() {
+        
+        close();
+
+        delete();
         
     }
